fix: Improve HPC error shutdown to improve logging [FE-44]
On Slurm job failures we have not reliably been getting the full logs
posted to the master.   When processing task-logging-teardown.sh
on error paths we commonly hit the 30sec timeout and therefore
exit the container which terminates all processing with SIGKILL.
When errors occur horovodrun/pid-server will terminate the top-level
children launched via SIGTERM, however, this only terminates the
top-level children when there are actually multiple layers.  When
only the top-level children are killed others still are holding
stdout/stderr active and therefore the log processing does not shut
down.

The change in this review is to enhance determined.utils.forward_signals
method to additionally forward the signal to the children of the specified
process in addtion to the process itself.