ci: track unreleased cherry-picked PRs during release
We've gotten the feedback that it would be nice to know what PRs have
been cherry-picked but are not yet available to test. Cherry-picked PRs
are now moved to a new unreleased status; adding tracking issues is
delayed until `track-pr release-unreleased-prs` is run (which happens
when an RC is cut). The PRs are then moved to a new released status
instead of being deleted. The process doesn't touch them after that, but
after a few releases I've found myself wanting to see what fixes have
been merged as part of the project view.