Merge data and row height menus in explist v2
This merges the data and row height menus together into one menu. The
icon used for the infinite scroll menu item is brought into the icon
component. The icon component is also refactored a little to ensure that
all svg icons are accounted for and won't overlap with the font icons.