ci: correctly change item status in release tracking
Previously, we were not correctly setting the status of tracked PRs in
release projects; a PR in a project has an item ID indicating its
presence in the project that is distinct from the ID of the PR
itself. This may have something to do with some PRs that were observed
ending up in entirely unexpected places, though the old code should have
just done nothing rather than moving things to wrong places, so I'm
still not sure what was going on there.