feat: Replace sum and count training metrics with mean in new experiment list
This replaces the sum and count metric columns for training metrics with
a `mean` column that returns the arithmetic mean of the training
metrics. The Columns API and experiment search api are updated with
support for the new column. Note that the validation metrics are
currently untouched because the column api and sort and filter
automatically choose between max and min based on the experiment config.