chore(actors): refactor allocation actor without actors (#7391)
This change migrates the allocation actor to an 'allocation service' instead. The service mostly manages allocation structs that look similar to allocation actors, but have a clearer API---and the service's API should be much clearer. It also gets rid of a few stray APIs I was able to prove we don't need anymore (e.g., GetContainerResourcesState, which was not serving the purpose from the original PR that it was added for and was no longer needed).

Ideally, this will be followed up by another refactor to move the allocation service into the rm code, but that will take longer.