fix: Improve HPC error shutdown to improve logging [FE-44] (#7488)
On Slurm job failures we do not reliably get the full logs
posted to the master from non-chief ranks.   When processing task-logging-teardown.sh
on error paths we commonly hit the 30sec timeout and therefore
end up with the entrypoint returning which which results in
termination of all processing with SIGKILL.
When errors occur pid-server will terminate the top-level
children launched via SIGTERM, however, this only terminates the
top-level children when there are actually multiple layers with
various process groups under the sshd.  When
only the top-level children are killed others still are holding
stdout/stderr active and therefore the log processing does not shut
down cleanly leading to potential message loss.

This change adds --signal-children support to the pid_server such that
whenever a signal is forwarded it is also forwarded to all children
launched under that the subprocess.  When deepspeed/horovod launch
the remote sshd, this option is
added so that on shutdown we ensure that children under the sshd are
shutdown as well as the sshd itself which enables logging to get
flushed properly.