build: convert svgs to react by default
This changes the default import behavior of svgs to convert the svg to a
react component. The previous default behavior can be accessed by adding
a `?url` query parameter to the import path, like so:

```tsx
import ScrollIconUrl from './icons/ScrollIcon.svg?url'
```

We planned on using svgr, but it turns out it has to use babel for the
conversion, so instead we're converting the parsed and cleaned svgs by
converting the ast to estree and modifying it directly there, then
outputting the resulting tree.