ci: handle duplicate cherry-picks of a PR to release branch (#7502)
Due to delays in running GitHub Actions, a PR that is labeled for
release cherry-picking immediately before or after being merged can go
through the cherry-pick code path twice. (The on-label action sees a
merged PR and the on-merge action sees a labeled one, so they both think
they have to perform the cherry-pick.) The second cherry-pick then fails
because Git sees that the commit would be empty and complains.

This change updates that code to check if a cherry-pick failure was due
to a would-be-empty commit and skips the cherry-pick in that case.

Ideally, we would fix the concurrency issue at the GitHub level, but
that sounds like more of a hassle. At an intermediate level of ideality,
we would tell `git cherry-pick` to automatically skip the cherry-pick
when it would be empty, but that doesn't seem to be
possible. (`--allow-empty` doesn't cover this case, while
`--keep-redundant-commits` would leave empty commits in the release
branch; that could then be fixed up after the fact, but at that point it
doesn't seem better than this approach.)