fix: remove selected file state from codeeditor
this fixes an issue where reloading the experiment code page didn't
result in the selected file in the view. The fix feels large, so here's
a walkthrough of what's going on:

1. codeviewer component only reads in the file path from props on mount,
but the experimentcodeviewer updates the file path as settings loads in
2. codeviewer component does this because otherwise the component would
have to render twice or load twice in order to load the files
3. codeview component has this problem because the selected file node
state is stored along with the loading state of the content

to fix this:

1. the selected file node and file loading state are decoupled. the selected
file node now is derived from the props, and the file loading state is
now changed as an effect of the file node changing.
2. Because of this, the `content` prop of the file node object has been
removed -- instead we always use the `get` function to get the file contents
3. Because the codeditor accidentally assumed async content was base64
encoded ascii, that functionality was moved to the api config