chore: Parse and format APIExceptions (#7531)
Previously, the body of an APIException was parsed as a JSON object and a message string was parsed from the result to be the exception's message. Currently, though:

- the only key present in an APIException's body is error (not message)
- the error value is a JSON-encoded dict, not a string.

This change now pulls out the "error" when possible and encodes the result in a formatted string that becomes the exception's message.  [MLG-454]