fix(actors): trial lifetime must contain allocation lifetime, still (#7615)
Introduction of the AllocationService changed ownership of an allocation from the
task that started it to the service. This causes issues because if the trial ends before
the allocation, the group may be unregistered by the experiment and cause RMs to
panic when trying to find the group associated with an RM scheduling.

(cherry picked from commit 3cb434de1beb8bee7a8e0c9ace0ef2e7720df997)