fix: Avoid potential NPE on delayed messages [FE-167] (#1002)
On Greenlake MLDEaaS deployment the file system used for job_storage_root
sometimes goes away for a bit.  This causes hangs, arbitrary launcher service
restarts, etc.   Occastionally, this causes dispatch monitoring events to be
delayed so long that other portions of the code have already detected the failure
and terminated the allocation.   When the DispatchStateChange eventually gets
delivered the task does not exist, and causes an NPE.   Ignore updates for
tasks that are no longer active.