chore(actors): refactor pod.go (#7617)
This refactors the pod actor to not be an actor. To do this, it was convenient to
move the request queue to use channels rather than multiple layers of callbacks.

As a dev note, there was a gnarly bug uncovered when making this change:
In k8s reattach tests, the sleep 10 commands don't actually die after 10 seconds
because FluentBit is stuck, up, retrying it's log shipping. They die as soon as you
bring the cluster back up, so reattach tests are testing that we don't miss
immediate state changes after recovering from medium-duration disconnects.
The old actor code relied on the state updates going through a few actors and
getting slowed down enough that the reattached resources got back to the
allocation first, this change affected the timings enough, `refreshPodState` was
introduced to fix it.