docs: Update the installation guide
this initial PR updates the installation guide and moves most of the old "basic setup" page into an "installation options" page. separately, we'll create an advanced setup checklist for each advanced option including on-prem, aws, gcp, k8s, and slurm. the main purpose is to provide users with an updated basic installation guide.