feat: streaming updates python client
This includes a new code generator.

Our existing code generation is proto -> openapi -> tweaks -> python,
with an alternate proto -> go path for the server code.

This is more direct, it's go -> python; the go code is the source of
truth and we can write the generator in go to use go's ast package.