fix: mitigate user settings race conditions
This mitigates an issue that appeaers on installations that have rbac
enabled. When updating settings, it's sometimes possible that a race
condition happens between the update request and the polled settings
request. becaue we optimistically update the local state, this can cause
the app to appear like it's not accepting the changes, only for it to
eventually settle after the next polling call.

We fix this issue by adding a grace period to the beginning of the
polling method. we keep track of all in-flight update calls and wait for
up to 500ms for those to complete before making the `getUserSetting`
call. If there are no updates in flight, the poll continues without waiting.