chore: Fix slurm determined-agent startup/cancel  [FE-165] (#1053)
The startup of the determined-agent on the Slurm cluster was passing
'localhost' to the agent when starting it.  It needed to use $HOSTNAME
so that the remote determined-agents started can connect back properly
from remote hosts.

Add a 15 min timeout to the job to avoid potential orphaned agents
on the cluster if the job is canceled or dies.

Set a unique job name so we can cancel just this job and not any others
that might potentially be running on the slurm cluster.