fix(allocation): allocation lifetimes should contain resource lifetimes (#7944)
We saw a NPE when the a scheduler group was removed before one of its 
allocations, since there is periodically code that wants the group for an 
allocation. Allocations should never outlive groups in the scheduler; the 
messages are sent in the right order but because the group message goes through 
from group actor->RP and the allocation message goes from allocation->RM->RP, 
if the RM is backed up the removals can happen out of order.

Since groups already wait on allocation exits, the fix just makes sure 
allocations wait for RMs to confirm receipt of the ResourcesReleased message 
before exiting.