chore: Improve failure diagnostics in shell test [FE-216] (#7932)
We are getting spurious failures of this test and it provide
no info when it fails.  On failure it also hangs until the
shell times out since the shell is still executing and therefore
doesn't close the stream we are reading.   Add an exit to force
an EOF so the test fails quickly on failure.