chore: Allow ROCm task env to use /tmp [FE-229] (#1063)
Our ROCm images generate temporary files in /tmp during the entrypoint
setup before this script is invoked. If the filesystem type is private
to the container ('overlayfs' with Singularity/Podman, or 'tmpfs' with enroot)
so we know we are not going to touch user files, then move the files to
the per-process local storage tmp we create, and the rm /tmp  below will
succeed so we can replace /tmp with a soft-link.