ci: more precisely select files for splitting in E2E tests
In ef59afa4b2fc, a pre-splitting selection step was added to work around
an unfortunate behavior of CircleCI's splitting tool when passed a file
with no tests to be run. The code was later amended to try to handle
slightly more complex marks, and a whole other thing was done on top of
that in EE, but really we should just use pytest itself to print out
exactly the right set of files without reimplementing any of its logic.

This change also adds code to defend against something going wrong and
causing no files to be selected, which would cause all tests to be run
in every split. (That's what happened in EE that caused this change
getting done.)