fix: fix sorting in GetUsers endpoint
This fixes an issue where adding a sort parameter in the `GetUsers`
endpoint would result in the endpoint throwing a 500 error. This is
because the order expression was invalid due to how bun escapes the
input to `Order`. We instead clean up the logic by always ordering by
the sort column and only adding the id sort if no other sort is available.