chore: Remove model_id from default constructors (#8015)
Usage of the default constructor is offered as "advanced" usage to let people call resource object
methods without first contacting master to get a fully instantiated object. In Model, this might
look like:

```
my_model = model.Model(name='foo', session=...)
my_version = my_model.get_version()
```

This works because the only resource-specific information `Model.get_version` uses from a Model
object is its name.

In commit #7922, we explicitly left in ModelVersion.model_id because there is method
(determined.core._experimental.report_task_using_model_version`) that takes a `ModelVersion` and
creates a `bindings.v1TrialSourceInfo` object using its `model_id`. But this is not the pattern the
partially-instantiated "advanced" usage is designed to support. We didn't need to leave `model_id`
in the constructor for this usage.

For `report_task_using_model_version`, a user needs to create a fully-hydrated object using one of
the existing `ModelVersion`-returning methods available to them. This is how its only reference in
the repo currently does it.

This change:

* removes `model_id` from `ModelVersion.__init__`
* removes `model_id` from `Model.__init__`
* no longer hydrates name during a `ModelVersion.reload` (to treat name
  like an id, not a mutable attribute)