chore: move ui kit to separate repo
This moves the source of the ui kit to a separate repo at
determined-ai/determined-ui. the old UI kit is removed and all imports
are updated. The new UI kit doesn't currently have CI or netlify set up (we
need to set up another set of github credentials so the differ works),
so the design kit and differ will continue to run here while that gets
up and going.