fix: hotfixes for ship_logs.py
Line endings were being dropped whenever the rank or level regexes
matched.  Fix includes a regression test.

Also k8s (and slurm, though that fix is on the EE side) was not using
ship_logs.py at all, as apparently the only place ToDockerSpec was being
used is in the agent resource manager.