fix: move couldn't-connect-to-master message into ship_logs.py (#8127)
A while ago, the first HTTP call we made to the master was in
prep_container.py.  There, we emitted a special, super-detailed message
to help users with broken container settings or proxy settings or
certificate settings to help sysadmins debug their clusters.

That message now belongs in ship_logs.py.  Also, we do a special
round-trip to the master just to check that connection, before we even
start a child process.