fix: remove 'contents' parameter from remove_notes
Project notes have name and contents. There is no uniqueness constraint
on name within a project. This gives an awkward interface for removing a
note by its name.

Previously, we had logic that optionally took a "contents", so a user
could narrow down _which_ note to delete when multiple notes had the
name as the note to be removed. But this interface is complicated, and
what's the likelihood that a user will event want it. Even more, we
hadn't implemented all the if/then casing right with "contents", and
this method had a bug as implemented.

This fixes the bug by removing the feature.