fix: prevent extra updates when observing settings store (#8212)
This fixes a class of issues that occur when migrating a component from
the `useSettings` hook to directly reading from/writing to the settings
store. Previously, the settings store would update the underlying state
map after each polling request. This causes issues for consumers that
have object values in their settings, because those objects would then
be replaced with identical objects after the settings poll finished.
When passing the object value into a hook dependency, the consumer then
updates along with the settings, even though the values don't change.
This can cause issues when the hook is responsible for things like
loading data, because it can cause extra calls and block user
interaction.

We fix this by doing a deep equality check on the settings after merging
with the polling results. If the two are equal, we discard the new
object and return the old one, preventing downstream updates.

(cherry picked from commit b6e28d7307e002de0c7057663d1fe06303062314)