chore: Update make slurmcluster to verify the status of the launcher (#1107)
We have seen at least one instance of the launcher service not being
reachable after a make slurmcluster in the nightly tests.   Assume
the most likely problem of a timing issue.   Check the status of the
launcher service (which verifies it is properly responding).  If this
is not the problem, it is likely the ssh tunnel spin-up time, but hopefully
this will be sufficient.