fix: pin ruamel.yaml<0.18.0
There was a breaking change in the ruamel.yaml API, which has the
unfortunate effect on our cli of causing commands like `det e config` to
emit absolutely no error message, just exit 1 silently.

The fix is too big to squeeze into the release so we will use a pin
until the real fix is ready.