fix: dsat with all yaml formats
Oops, we forgot to set the "make it work" switch to the "on" position.

For a valid yaml string:

    s = "image: {gpu: image:tag}"

This innocent looking code does not work:

    yaml.YAML(typ="safe").load(s)

But this code does work:

    yaml.YAML(typ="safe", pure=True).load(s)

Setting pure=True cause ruamel.yaml to skip the C implementation, which
is apparently not entirely correct.  Why not?  Who knows.

This was exposed by the recent conversion of some
`yaml.dump(config)` calls to `util.yaml_safe_dump(config)` in the e2e
experiment utils.