fix: prevent settings store from triggering rerenders on poll
this fixes an issue where components that used the settings store would
rerender on poll even if the polling function returned identical data.
This is slightly different from #8212 -- there, the entire settings
object was recreated, causing anything that relied on objects within the
settings to refresh. This fixes the issue where each render returned the
same data in a new Loadable.