chore: deployment's --dry-run option doesn't print template (#8303)
Pre-patch, --dry-run:
1. validates flags
2. checks quotas (unless otherwise suppressed)
3. prints a CF template

This patch removes the 3rd step.

It was easy to confuse the template printed by --dry-run with the
parameters that would be actually used when deploying a cluster
(https://github.com/determined-ai/determined/issues/7978 for one such
bug report).

It's tempting to remove the --dry-run flag entirely. This patch takes
the stance that it's helpful to have an option that at least validates
flags without doing something that's hard to undo, and so even a
restricted --dry-run still has some use.