remove fit-content from timeago
this cleans up the `TimeAgo` css. previously, `TimeAgo` rendered as a
div with width set to fit-content in #5149. This was to fix the div
stretching to fill its container as a block-level element. Now, the text
renders as a span, so the width of the container fits its content
naturally. While in there, we remove the classname prop, which was only
used in one place in the codebase.