fix(tests): lower e2e_gpu_quarantine parallelism (#8363)
e2e_gpu_quarantine is failing on main because one of the splits has no
tests, but there is only one test to run so we don't need splits at all. I
could fix the shell to not panic if we run a useless split, but I think it
should still, honestly. That's a lot of wasted resources.