fix: don't dup logs on singluarity and podman (#1130)
The old log shippper used strtobool and so a "=False" env var would
evaluate to False.  The new log shipper treats non-empty env vars as
True.  This was the source of the bug.

However, I see no reason why we can't let the user's setting for
DET_EMIT_STDOUT_LOGS be definitive.  I don't know why they would set it
here but it feels like it should be their prerogative.