migrate detaileduser and experimentitem types to io-ts
This migrates the detaileduser and experimentitem types to io-ts. This
is because these types are in use as codecs for the immutableobservables
in the user and experiment stores respectively. To prevent the codec
type from drifting away from the concrete type, we instead derive
everything from the codec type. A small knock-on effect is that the
helper `ValueOf` type and codec have been moved to a separate module to
prevent circular imports.