fix(rm): tasks shouldn't hang on restore failures (#8486)
We had a few bugs, but the main one was that on receipt of a ResourcesRestoreFailure allocations
would not exit but just chill waiting on a ResourcesReleased event that would never come (RMs have never
sent this). The other issue was the RM was relying on allocations to cleanup their resources but if restore failed
they wouldn't be around to do this, so now we send an extra SIGKILL from the RM just in case.

I don't think these fixes are 100% optimal, I think there is some architectural change that solve all these and
similar issues, but I haven't dreamed up the best approach yet.

I also fixed up some logging and type names to make debugging and readability better.