fix: ensure filter columns are valid when selecting special columns (#8517)
this fixes an issue in the experiment list where selecting a column that
has specific values after selecting a column that does not causes the
resulting query to be invalid. this is because the code intended to do
so checked if the previous value was special, not if the value the field
was turning to.