ci: unit-test login scenarios, and others (#8471)
The login scenarios unit test replaces two e2e tests, and then provides
even more test coverage beyond those.

The intention of this test is more "clearly document the current
behavior" rather than "enforce the current behavior forever".

There are decisions we've made, especially regarding the default login
credentials, that feel suspect to me.  I hope this test is useful for
highlighting the effects of changing any of those decisions, so we can
make those changes in an educated way.  I do not intend for this test
to force us to keep the exact implementation we have forever.

Additional tests here are focused on enabling future simplifications of
e2e tests, which are currently providing an unhealthy amount of cli test
coverage