fix(tasks): persist rendezvous readiness (#8545)
We had a bug where a user tried to pause an experiment but it unexpectedly was killed. The root
cause was that we don't persist the ready bit, so it is lost on a master crash, and the ready bit is what
we check to decide if we should bother with graceful termination at all.