fix: prevent theme settings conflict in multi-client/window scenario
This fixes an issue where the user management table appeared to refresh
constantly. the user management page is set up to refresh the
table when its underlying persisted settings refresh, which can happen
inadvertently when multiple clients are using the same user account. The
theme setting was set up to always update the settings table to the
local value on conflict, which causes a conflict when one client changes
the theme. To fix this, we:

1) ensure that the user settings store does not trigger refreshes when
setting to an identical value. note that we aren't able to use the deep
or immutable observable wrappers because the value is an immutable
collection wrapped in a plain object -- immutable collections are
stateful, so a simple deep check wouldn't work, and because the top
value isn't an immutable collection, the immutable observer wouldn't
work.

2) move themesettings out of usesettings. this prevents the
updateSettings fucntion from being marked as an effect dependency

3) ensure that the remote wins when the theme is in conflict