chore: cli describe workspaces no longer also lists projects
Old:
`det workspace describe <workspace>` (when executed with out `--json`)
used to first print a table of information about the passed workspace
and then would print another table listing all the projects that belong
to the workspace.

New:
`det workspace describe <workspace>` *only* describes the passed
workspace

Justification:
`det workspace list-projects <workspace>` is already a thing. If a user
wants to list workspaces they can run that command. The old behavior
was awkward to implement because it needed to assume parameters to
list-projects that would normally be passed as flags to `det workspace
-list-projects`.