ci: entirely remove mmdetection test runs
All of the tests are now marked as skipped in the Python code, so the
nightly run was failing and there's no reason to run it anyway.

We could tweak the test splitting code so that it doesn't fail if there
are no tests, but I would prefer not to unless there are plans to
reenable mmdetection tests fairly soon.