docs: Add section on viewing topology
## Description

TECHWR-369

As an MLE, I would like to have a macro understanding of how the GPUs and nodes in my cluster are distributed within Determined and which slots on which GPUs are occupied, enabling me to know if my job will run and/or if there are sufficient resources for it to do so.

To visualize each node and the number of slots available and which slots are active vs used, visit the Topology section in the resource pools' details page.