fix: stop using distutils (for python 3.12) [MLG-1519]
distutils is removed in python 3.12.  We shouldn't really be using it
anyway, but rather than do a breaking change to switch to the behavior I
want us to have, here's a quick fix that unbreaks us, and I made a
ticket to do the breaking change at some point in the future.

Note that to manifest the failure, you'll need to uninstall setuptools
from your python environment (which provides its own distutils package
starting in python 3.12).